refactor: refactoring sub commands and rejigging tests as required for --about and --example commands
