+++ cli/main.py
@@ -4,13 +4,21 @@ import shutil
 
 from runner import *
 
-CONTEXT_SRC_FILE = os.path.dirname(__file__) + '/../init/context.py'
+TEST_SRC_FILE = os.path.dirname(__file__) + '/../init_example/test_fizzbuzz.py'
 TEST_DST_ROOT = './tests'
+
+TEST_LOGIC_SRC_FILE = os.path.dirname(__file__) + '/../init_example/fizzbuzz.py'
+TEST_LOGIC_DST_ROOT = '.'
+
+CONTEXT_SRC_FILE = os.path.dirname(__file__) + '/../init_example/context.py'
+CONTEXT_SRC_FILE = os.path.dirname(__file__) + '/../init/context.py'
+
 SNAKE_EMOJI = '\U0001F40D'
 
 @click.group(invoke_without_command=True)
 @click.pass_context
 
+### Click flags
 @click.version_option(
     version=None,
     help='Display the current version.',
@@ -32,30 +40,38 @@ SNAKE_EMOJI = '\U0001F40D'
     help='Ininitalise your project with an example fizzbuzz project',
     is_flag=True
 )
-
+### Click argument
 @click.argument('filepath', required=False)
 
 def cli(ctx, filepath, init, about, example):
 
-    # --init
+    # Flag --init command
     if ctx.invoked_subcommand is None and init:
         create_test_dir()
         create_context_file()
 
+    # Flag --about command
     elif ctx.invoked_subcommand is None and about:
-        print_about()
+        print_about_message()
 
+    # Flag --example command
     elif ctx.invoked_subcommand is None and example:
-        click.echo('hello')
+        click.echo("Initialising fizzbuzz example")
+        create_test_dir()
+        create_fizzbuzz_file()
+        create_fizzbuzz_test_file()
+        create_fizzbuzz_context_file()
 
+    # Flag serpent <filename> argument command
     elif ctx.invoked_subcommand is None and filepath:
         if filepath:
             run_specific_file(filepath)
 
-    # --run all tests
+    # serpent command
     elif ctx.invoked_subcommand is None:
         run_all()
 
+## functions
 def create_test_dir():
     if not os.path.exists('tests'):
         os.makedirs('tests')
@@ -71,5 +87,17 @@ def run_specific_file(filepath):
     click.echo("running specfic test file")
     run_test(filepath)
 
-def print_about():
+def print_about_message():
     print(f"Made with {SNAKE_EMOJI} by Alexandra McCarroll, Tom Betts, Richard Hewitt, Hemesh Unka (February 2018 Cohort - Makers Academy)")
+
+def create_fizzbuzz_test_file():
+    if not os.path.isfile('tests/test_logic.py'):
+        shutil.copy(TEST_SRC_FILE, TEST_DST_ROOT)
+
+def create_fizzbuzz_context_file():
+    if not os.path.isfile('tests/context.py'):
+        shutil.copy(CONTEXT_SRC_FILE, TEST_DST_ROOT)
+
+def create_fizzbuzz_file():
+    if not os.path.isfile('logic.py'):
+        shutil.copy(TEST_LOGIC_SRC_FILE, TEST_LOGIC_DST_ROOT)
